steamworks-map-adder/requirements.txt
PyYAML===6.0.2